+++ Model/DropReference.cs
@@ -1,8 +1,11 @@
-﻿namespace lib.remnant2.analyzer.Model;
+﻿using System.Diagnostics;
 
+namespace lib.remnant2.analyzer.Model;
+
+[DebuggerDisplay("{Name}")]
 public class DropReference
 {
     public required string Name { get; set; }
-    public bool IsDeleted { get; set; } // This item is already looted or quest completed, etc
+    public bool IsLooted { get; set; } // This item is already looted or quest completed, etc
     public required List<string> Related { get; set; }
 }