+++ Model/Location.cs
@@ -2,19 +2,22 @@
 
 namespace lib.remnant2.analyzer.Model;
 
+[DebuggerDisplay("{Name}")]
 public class Location
 {
     public required string Name;
     public required List<string> WorldStones;
     public required List<string> Connections;
     public bool TraitBook;
-    public bool TraitBookDeleted;
+    public bool TraitBookLooted;
     public bool Simulacrum;
-    public bool SimulacrumDeleted;
+    public bool SimulacrumLooted;
     public required List<DropReference> WorldDrops;
     public required List<DropReference> DropReferences;
     public required string Category;
     public required List<LootGroup> LootGroups;
+    public required List<LootedMarker> LootedMarkers;
+
 
     public bool Bloodmoon
     {
@@ -97,6 +100,7 @@ public class Location
         WorldDrops = [],
         DropReferences = [],
         Category = "Ward 13",
-        LootGroups = []
+        LootGroups = [],
+        LootedMarkers = []
     };
 }