+++ Model/RolledWorld.cs
@@ -1,4 +1,6 @@
-﻿namespace lib.remnant2.analyzer.Model;
+﻿using lib.remnant2.saves.Model;
+
+namespace lib.remnant2.analyzer.Model;
 
 // Represents part of the data from a single save_N.sav: either adventure data or campaign data
 public class RolledWorld
@@ -12,8 +14,25 @@ public class RolledWorld
     public Zone Ward13;
     public List<Zone> AllZones => [ Ward13,..Zones ];
 
+    private Character? _character;
+    public Character Character
+    {
+        get => _character ?? throw new InvalidOperationException("Character is not set for RolledWorld, this is unexpected");
+        set => _character = value;
+    }
+
+    public required string Difficulty;
+    public TimeSpan? Playtime;
+
     public bool CanGetItem(string item)
     {
         return AllZones.Any(x => x.CanGetItem(item));
+    }
+
+    public bool IsCampaign() => Zones.Exists(x => x.Name == "The Labyrinth");
+
+    public bool CanGetAccountAward(string award)
+    {
+        return false;
     }
 }