+++ Model/SaveSlot.cs
@@ -7,4 +7,6 @@ public class SaveSlot
     public RolledWorld? Adventure;
     public required List<string> QuestCompletedLog;
     public bool HasTree;
+    public required List<LootItem> CassShop;
+    public TimeSpan? Playtime;
 }